backend/requirements.txt
Django>=5.1,<6.0
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]==5.3.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0
psycopg[binary]>=3.1.0
python-decouple==3.8
Pillow>=10.3.0
gunicorn==21.2.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
